+++ CHANGELOG.md
@@ -7,6 +7,9 @@
 - Baywatch default parameters added: `CB_BAYWATCH_ENABLED` and`CB_BAYWATCH_EXTERN_LOCATION`
 - Logs are saved via lospout
 - TLS client certificate needed by Cloudbreak is generated with `cbd generate`
+- Command `aws delete-role ` added
+- Command `aws generate-role` added
+- Command `aws show-role` added
 
 ### Removed