+++ CHANGELOG.md
@@ -21,6 +21,7 @@
 ### Changed
 
 - AWS authentication env varibale is fixed to use the correct AWS_SECRET_ACCESS_KEY (instead the old AWS_SECRET_KEY)
+- Using sequenceiq/ambassadord:0.5.0 docker image instead of progrium/ambassadord:latest
 
 ## [v0.5.3] - 2015-06-03